requirements.txt
flake8
wemake-python-styleguide
flake8-pytest-style
flake8-django
flake8-logging-format

Django
graphene==2.1.8
graphene-django==2.15.0
django-graphql-jwt==0.3.0
PyJWT==1.7.0
python-decouple==3.4

